ml/requirements.txt
# Machine Learning Dependencies
numpy>=1.26.0,<2.0.0
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1
joblib>=1.3.2

# API Framework
flask==2.3.3
flask-cors==4.0.0

# Database connection
psycopg2-binary==2.9.7
python-dotenv==1.0.0

# Visualization (optional)
matplotlib==3.7.2
seaborn==0.12.2

# Advanced ML models (optional)
xgboost==2.0.0
lightgbm==4.0.0

# Fuzzy Logic
scikit-fuzzy==0.4.2

